mortgage-core/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers
psycopg2-binary
requests==2.31.0
python-dotenv
dj-database-url